Fifa Ultimate Team Sniper Bot
Software made in C# with a goal of auto buying a specific player with a given cost (cheaper) and sell it (more expensive) to make in game money.

The program uses CefSharp as a web browser to open the web version of the FUT Web App, then loging on given account.

The bot itself navigate between differents pages and click on specics buttons thanks to javascript functions that simulates touches (or mouse click) on the web page.

The auto-buy function search and refresh the transfer market until a card is available at the given price. When a card is available, the bot will buy it faster than a real human player (with random delay, to avoid bans from EA) and add it to the transfer list.